server/requirements.txt
# Core Django
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5

# Environment Management
python-decouple==3.8
python-dotenv==1.0.0

# Email
django-ses==3.5.2

# Validation
email-validator==2.1.0

# CORS
django-cors-headers==4.3.1

# Filtering & Search
django-filter==23.5

# Image Processing
Pillow==10.1.0

# API Documentation
drf-yasg==1.21.7

# Production Server
gunicorn==21.2.0

# Monitoring
sentry-sdk==1.39.2